+++ lib/components/DataGridVue.vue
@@ -38,6 +38,8 @@
             @keydown.space="toggleFilterOptionsShown"
             @keydown.enter="toggleFilterOptionsShown"
             :aria-label="`Filter options are currently ${filterOptionsShown ? 'shown' : 'hidden'}, use space bar to toggle visibility.`"
+            role="checkbox"
+            :aria-checked="filterOptionsShown"
           >
             <Icon name="filter" />
             <span>{{ filterOptionsShown ? 'Hide' : 'Show' }} Filter Options</span>
@@ -61,6 +63,7 @@
             :aria-label="
               filterSummary ? `Clear all filters. Currently filtering by ${filterSummary}` : `Clear all filters. No filters currently applied.`
             "
+            role="button"
           >
             <Icon name="clear-filter" />
             <span>Clear Filters</span>
@@ -79,6 +82,8 @@
             @keydown.space="toggleColumnSelectionShown($event)"
             @keydown.enter="toggleColumnSelectionShown($event)"
             aria-label="Open the add/remove columns menu."
+            role="checkbox"
+            :aria-checked="!!popupOptions"
           >
             <Icon name="add-column" />
             <span>Add/Remove Columns</span>