+++ vuepress/shared/changelog.md
@@ -4,8 +4,8 @@
 
 ### v3.0.1
   - bugfix: header title and aria label incorrectly informs user that a column can be sorted when sort is turned on for the grid but not that specific column
-  - accessibility: scrollable region must have keyboard access for dgv-data-grid-body
-  - accessibility: aria-label attribute cannot be used on a span with no valid role attribute dgv-filter-operator-role
+  - accessibility: scrollable region must have keyboard access for `dgv-data-grid-body`
+  - accessibility: `aria-label` attribute cannot be used on a span with no valid `role` attribute for `dgv-filter-operator` and `options-header`.
 
 
 ### v3.0.0