+++ vuepress/shared/changelog.md
@@ -6,6 +6,7 @@
   - bugfix: header title and aria label incorrectly informs user that a column can be sorted when sort is turned on for the grid but not that specific column
   - accessibility: scrollable region must have keyboard access for `dgv-data-grid-body`
   - accessibility: `aria-label` attribute cannot be used on a span with no valid `role` attribute for `dgv-filter-operator` and `options-header`.
+  - docs: fix critical accessibility errors reported by axe
 
 
 ### v3.0.0